backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic>=2.7.0
pydantic-settings>=2.1.0
pydantic[email]>=2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
anthropic==0.39.0
google-genai>=0.3.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
